refactor: move the adjustFilePath compat delegates into path-resolver helpers
