+++ src/utils/path-resolver/helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './adjust-file-path';
+export * from './adjust-file-paths';
