+++ src/utils/path-resolver/index.ts
@@ -1,3 +1,4 @@
 export * from './constants';
+export * from './helpers';
 export * from './module';
 export * from './type';
